Customized Aurna-Lightbox
Aurna Lightbox is a javascript website plugin made by me to make an Iframe Lightbox for ajax or dynamic websites. I customized the code a little and made the iframe to render a div. Soon I'll be releasing a stable version of Aurna Lightbox. It's simple, raw javascript and CSS and tiny in size.