Tampermonkey Script for UIU UCam
A rough tampermonkey script for UIU Ucam, still working on it. 
The Routine Catcher is not done yet. Hope I'll be able to parse the Class routine from UCAM and comparatively will be able to show the filtered exam routine from a JSON file which contains the EXCEL file data of the exam routine of each trimester of UIU.